Fix for job failure plugins being run twice
After adding failure_hook handling to the Job.fail method, which is called by
the worker upon failure, it is no longer necessary for the Job to call it's own
failure hooks. Currently the failure hooks are being called twice. Once by the
job itself, and then the error is rethrown, where the worker catches it and calls
Job.fail where the hooks are run again.

This patch simply removes the large begin/rescue block inside of the job.